requirements.txt
altair
matplotlib
numpy
pandas
plotly
pydeck
country-converter
scikit-learn
sqlalchemy
streamlit
umap-learn
hdbscan
optuna
seaborn